get CI green after ECS migration finish
Test compile fixes and handler guards to make `dotnet test` pass end-to-end.

src/ changes:
- CharNew/CharNewJob/CharRen/CharacterDelete packet handlers: early-return
  when `clientSession.HasSelectedCharacter` (was relying on removed
  `Session.SetCharacterAsync(null)` in tests).
- EntryPointPacketHandler: iterate `CharacterDto` directly instead of
  `Adapt<Character>()` (Character requires constructor args).
- CharNewPacketHandler: use `InventoryService` directly instead of mapping
  to legacy `Character` for starter items.
- SaveService: sync component-backed scalars (Hp/Mp/Level/…/Gold/…/Pos)
  into CharacterDto before persisting.
- MapWorld/MapInstance/MinilandService: `Kick` -> awaited `KickAsync`
  (fire-and-forget ChangeMapAsync was racing test asserts).
- IMapInstanceGeneratorService.RemoveMap -> RemoveMapAsync; callers await.
- Shop/Battle/Movement/Game/UseSkill handlers: replace throwing
  `GetCharacter` with `TryGetCharacter` out-pattern.
- PlayerBundleExtensions.Add(Additional)SpPoints: drop stray `(byte)` cast
  that truncated SP cap to 255.
- Remove unused `LogLanguageKey.CHARACTER_NOT_INIT` + its resx entries.

test/ changes:
- Migrate GameObject + PacketHandlers tests off the legacy `Character`
  class and `Session.SetCharacterAsync`. Use `ClearPlayerEntity`,
  captured `CharacterDto`, and direct `ClientSession`/`PlayerComponentBundle`
  mocking for `ISessionRegistry`/`ISaveService`.
- Expose `MapChangeService` on `TestHelpers` and use the shared
  `SessionGroupFactory` mock so ChangeMapInstanceAsync works in tests.
- Stub `IMinilandService.GetMinilandPortals` to return an empty list so
  the map-change flow doesn't NRE.

Result: 0 warnings, 0 errors; 820 tests passing across all projects.

+++ src/NosCore.GameObject/Ecs/Extensions/PlayerBundleExtensions.cs
@@ -764,15 +764,15 @@ public static class PlayerBundleExtensions
     public static void AddSpPoints(this PlayerComponentBundle player, int spPointToAdd, IOptions<WorldConfiguration> worldConfiguration)
     {
         var newValue = player.SpPoint + spPointToAdd;
-        player.SpPoint = (byte)(newValue > worldConfiguration.Value.MaxSpPoints
-            ? worldConfiguration.Value.MaxSpPoints : newValue);
+        player.SpPoint = newValue > worldConfiguration.Value.MaxSpPoints
+            ? worldConfiguration.Value.MaxSpPoints : newValue;
     }
 
     public static void AddAdditionalSpPoints(this PlayerComponentBundle player, int spPointToAdd, IOptions<WorldConfiguration> worldConfiguration)
     {
         var newValue = player.SpAdditionPoint + spPointToAdd;
-        player.SpAdditionPoint = (byte)(newValue > worldConfiguration.Value.MaxAdditionalSpPoints
-            ? worldConfiguration.Value.MaxAdditionalSpPoints : newValue);
+        player.SpAdditionPoint = newValue > worldConfiguration.Value.MaxAdditionalSpPoints
+            ? worldConfiguration.Value.MaxAdditionalSpPoints : newValue;
     }
 
     public static void RemoveGold(this PlayerComponentBundle player, long gold)
